+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.0.0] - 2026-06-05
 ### Changed
 - Add interactor param into IInteractable.Interact function
 - Rename InteractableUnityEvent -> InteractableHandler
@@ -60,7 +62,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/1mbitshorde/InteractableSystem/compare/3.2.0...main
+[Unreleased]: https://github.com/1mbitshorde/InteractableSystem/compare/4.0.0...main
+[4.0.0]: https://github.com/1mbitshorde/InteractableSystem/tree/4.0.0/
 [3.2.0]: https://github.com/1mbitshorde/InteractableSystem/tree/3.2.0/
 [3.1.1]: https://github.com/1mbitshorde/InteractableSystem/tree/3.1.1/
 [3.1.0]: https://github.com/1mbitshorde/InteractableSystem/tree/3.1.0/